canbus: fix compile warning.

+++ modules/canbus/vehicle/lincoln/protocol/gear_66.cc
@@ -74,7 +74,7 @@ Gear66 *Gear66::set_gear_low() {
 void Gear66::set_gear_p(uint8_t *data, int32_t gear) {
   gear = ProtocolData::BoundedValue(0, 5, gear);
   Byte frame(data);
-  frame.set_value(gear, 0, 3);
+  frame.set_value(static_cast<uint8_t>(gear), 0, 3);
 }
 
 void Gear66::set_clear_driver_override_flag_p(uint8_t *bytes) {
